Fix error in filter generation's "interesting lists" logic
The non-archival settings for things that go to interesting lists had
a problem: it was selecting only the lists that you *weren't*
interested in, to treat as interesting lists. This caused the
generated filters to treat everything as an interesting email,
effectively archiving nothing.

This change makes it invert that selection to achieve the code's
stated goals.